backend/requirements.txt
fastapi[all]
sqlalchemy
pydantic
uvicorn
pytest
python-jose[cryptography]
passlib[bcrypt]
# Pin bcrypt to a version compatible with passlib's detection (>=3.2,<4 where long-password check doesn't raise)
bcrypt==3.2.2
